requirements.txt
celery==4.2.1
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.2.0
coreapi==2.3.3
PyYAML==3.12
psycopg2-binary==2.7.5